% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crmd.R
\name{crmd}
\alias{crmd}
\title{Convert Jupyter Notebook to R Markdown}
\usage{
crmd(input, path = getwd(), output = "convert_rmd.Rmd", extract = c("all", "code", "markdown"))
}
\arguments{
\item{input}{Path to the input `.ipynb` file.}

\item{path}{Directory where the `.Rmd` will be saved. Defaults to the current working directory.}

\item{output}{File name for the `.Rmd` file. Defaults to `"convert_rmd.Rmd"`.}

\item{extract}{One of `"all"`, `"code"`, or `"markdown"`. Controls which types of cells to extract.}
}
\value{
A message string: `"Convert success."`
}
\description{
Convert Jupyter Notebook (.ipynb) to R Markdown (.Rmd)
}
\details{
This function converts a Jupyter Notebook (`.ipynb`) file into an R Markdown (`.Rmd`) file.
It allows you to extract code cells, markdown cells, or both, and lets you specify
the output directory and filename.
}
\examples{
\dontrun{
crmd("notebook.ipynb")
crmd("notebook.ipynb", extract = "code")
crmd("notebook.ipynb", path = "reports/", output = "my_report.Rmd")
}
}

\author{
Jung-In Seo \cr
Jeong Wook Lee \email{gunzion12@gmail.com}
}

