% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crmd.R
\name{crmd}
\alias{crmd}
\title{Convert Jupyter Notebook (.ipynb) to R Markdown (.Rmd)}
\arguments{
\item{input}{Character scalar. Path to the input `.ipynb` file.}

\item{path}{Character scalar. Directory where the `.Rmd` will be saved.
Defaults to the current working directory.}

\item{output}{Character scalar. File name for the output `.Rmd` file.
Defaults to `"convert_rmd.Rmd"`.}

\item{extract}{Character scalar. Cell types to extract:
`"all"` (default), `"code"`, or `"markdown"`.}
}
\value{
(invisible) A message string: `"Convert success."`
}
\description{
Convert a Jupyter Notebook (`.ipynb`) into an R Markdown (`.Rmd`) file by
extracting code cells, markdown cells, or both. The result is a plain Rmd file
that can be rendered or edited in RStudio.
}
\details{
This function reads a Jupyter `.ipynb` file and transforms it into an `.Rmd`
document, maintaining the notebook structure. You can control which cells are
included using the `extract` argument. By default, both code and markdown
cells are included.
}
\examples{
\dontrun{
crmd("notebook.ipynb")
crmd("notebook.ipynb", extract = "code")
crmd("notebook.ipynb", path = "reports/", output = "lecture.Rmd")
}

}
\author{
Jung-In Seo \cr
Jeong Wook Lee \email{gunzion12@gmail.com}

Convert Jupyter Notebook (.ipynb) to R Markdown (.Rmd)
...
}
